Add files via upload
This application has been developed to calculate the estimated calories user will burn while biking from current place to destination. More to it, the user will get to know the pathway to reach the destination place and he/she can also update his/her information and an information of Bike.